申请加入

+++ cmd/server/v1/domain/plan.go
@@ -7,7 +7,7 @@ import (
 type PlanReqPaginate struct {
 	_domain.PaginateReq
 
-	CategoryId uint   `json:"categoryId"`
+	CategoryId int64  `json:"categoryId"`
 	Keywords   string `json:"keywords"`
 	Enabled    string `json:"enabled"`
 }
