+++ cmd/server/v1/domain/scenario.go
@@ -10,7 +10,7 @@ import (
 type ScenarioReqPaginate struct {
 	_domain.PaginateReq
 
-	CategoryId uint   `json:"categoryId"`
+	CategoryId int64  `json:"categoryId"`
 	Keywords   string `json:"keywords"`
 	Enabled    string `json:"enabled"`
 }
